organizando pastas da autenticação
