+++ application/apps/api/src/auth/auth.service.ts
@@ -53,6 +53,9 @@ export class AuthService {
 
     const payload = { sub: user.id, email: user.email };
 
-    return { access_token: await this.jwtService.signAsync(payload) };
+    return { 
+      access_token: await this.jwtService.signAsync(payload),
+      user_id: user.id // Incluindo o ID do usuário no objeto retornado
+    };
   }
 }
