+++ application/apps/client/src/components/LoginPage/LoginPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useForm } from "react-hook-form";
-import { useAuth } from "../../contexts/AuthContext";
+import { useAuth } from "../../contexts/auth/AuthContext";
 import { useNavigate } from "react-router-dom";
 import './LoginPage.css'
 
@@ -9,7 +9,7 @@ type LoginForm = {
   password: string;
 };
 
-const LoginPage = () => {
+export const LoginPage = () => {
 
   const { signIn } = useAuth();
   const { register, handleSubmit } = useForm<LoginForm>();
@@ -69,5 +69,3 @@ const LoginPage = () => {
     </>
   );
 };
-
-export default LoginPage;
