--- application/apps/client/src/contexts/services/api.ts
@@ -1,8 +1,0 @@
-// api.ts
-import axios from "axios";
-
-export const api = axios.create({
-  baseURL: 'http://localhost:3000', // Coloque a URL correta do seu backend
-});
-
-
