+++ application/apps/client/src/main.tsx
@@ -1,8 +1,7 @@
-import React from 'react'
 import ReactDOM from 'react-dom/client'
 import {App} from './App.tsx'
 import { BrowserRouter } from 'react-router-dom'
-import { AuthProvider } from './contexts/AuthContext.tsx'
+import { AuthProvider } from './contexts/auth/AuthProvider.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <AuthProvider>
