+++ application/apps/client/src/types/AuthContext.ts
@@ -1,0 +1,8 @@
+import { User } from "./User";
+
+export type AuthContextType = {
+    user: User | null;
+    signIn: (email: string, password: string) => Promise<void>;
+    signOut: () => void;
+    isAuthenticated: boolean;
+  };
