+++ application/apps/client/src/types/User.ts
@@ -1,0 +1,3 @@
+export type User = {
+    id: number;
+  };
